Refactor FIT export logic and cleanup types
