+++ utils/download.ts
@@ -1,0 +1,15 @@
+/**
+ * Downloads a Blob as a file in the browser.
+ * @param blob The Blob to download.
+ * @param filename The name of the file to save.
+ */
+export const downloadBlob = (blob: Blob, filename: string): void => {
+  const url = window.URL.createObjectURL(blob)
+  const a = document.createElement('a')
+  a.href = url
+  a.download = filename
+  document.body.appendChild(a)
+  a.click()
+  window.URL.revokeObjectURL(url)
+  document.body.removeChild(a)
+}
